+++ pkg/pillar/utils/wait/waitfor.go
@@ -8,7 +8,6 @@ import (
 
 	info "github.com/lf-edge/eve-api/go/info"
 	"github.com/lf-edge/eve/pkg/pillar/base"
-	"github.com/lf-edge/eve/pkg/pillar/containerd"
 	"github.com/lf-edge/eve/pkg/pillar/pubsub"
 	"github.com/lf-edge/eve/pkg/pillar/types"
 	uuid "github.com/satori/go.uuid"
@@ -145,30 +144,4 @@ func handleOnboardStatusImpl(ctxArg interface{}, key string,
 		return
 	}
 	ctx.Initialized = true
-}
-
-// WaitForUserContainerd waits until user containerd started
-func WaitForUserContainerd(ps *pubsub.PubSub, log *base.LogObject, agentName string, warningTime, errorTime time.Duration) error {
-	stillRunning := time.NewTicker(25 * time.Second)
-	ps.StillRunning(agentName, warningTime, errorTime)
-	checkTicker := time.NewTicker(5 * time.Second)
-	initialized := false
-
-	for !initialized {
-		log.Noticeln("Waiting for user containerd socket initialized")
-		select {
-		case <-checkTicker.C:
-			ctrdClient, err := containerd.NewContainerdClient(true)
-			if err != nil {
-				log.Tracef("user containerd not ready: %v", err)
-				continue
-			}
-			_ = ctrdClient.CloseClient()
-			initialized = true
-		case <-stillRunning.C:
-		}
-		ps.StillRunning(agentName, warningTime, errorTime)
-	}
-	stillRunning.Stop()
-	return nil
 }
